Agregando Ejercicio45 Tarea

+++ Ejercicios/Ejercicio45-Punto-de-venta-parte-VI/factura.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 double Subtotal, Total;
 string listaProductos;
-int Impuesto;
+double Impuesto;
